// src/types/athlete.ts
export type Athlete = {
    id: number;
    athleteCode: string;
    firstName: string;
    lastName: string;
    gender: string;
    age: number;
    dateOfBirth: string;
    country: string;
    sport: string;
    position: string;
    team: string;
    ranking: number;
    medals: number;
    matchesPlayed: number;
    wins: number;
    losses: number;
    winRate: number;
    heightCm: number;
    weightKg: number;
    yearsPro: number;
    salaryUsd: number;
    isOlympian: boolean;
    status: string;
    lastUpdated: string;
};